+++ config/config.go
@@ -20,6 +20,7 @@ const (
 	serviceDBVersion   = "v3"
 	executionDBVersion = "v2"
 	instanceDBVersion  = "v1"
+	workflowDBVersion  = "v1"
 )
 
 var (
@@ -46,6 +47,7 @@ type Config struct {
 		ServiceRelativePath   string
 		InstanceRelativePath  string
 		ExecutionRelativePath string
+		WorkflowRelativePath  string
 	}
 
 	Tendermint struct {
@@ -77,6 +79,7 @@ func New() (*Config, error) {
 	c.Database.ServiceRelativePath = filepath.Join("database", "services", serviceDBVersion)
 	c.Database.InstanceRelativePath = filepath.Join("database", "instance", instanceDBVersion)
 	c.Database.ExecutionRelativePath = filepath.Join("database", "executions", executionDBVersion)
+	c.Database.WorkflowRelativePath = filepath.Join("database", "workflows", workflowDBVersion)
 	c.Tendermint.Path = "tendermint"
 	c.setupServices()
 	return &c, nil