+++ execution/execution.go
@@ -35,6 +35,7 @@ func (s Status) String() (r string) {
 // Execution stores all information about executions.
 type Execution struct {
 	Hash         hash.Hash              `hash:"-"`
+	WorkflowHash hash.Hash              `hash:"name:workflowHash"`
 	ParentHash   hash.Hash              `hash:"name:parentHash"`
 	EventHash    hash.Hash              `hash:"name:eventHash"`
 	Status       Status                 `hash:"-"`
@@ -47,8 +48,9 @@ type Execution struct {
 }
 
 // New returns a new execution. It returns an error if inputs are invalid.
-func New(instanceHash, parentHash, eventHash hash.Hash, taskKey string, inputs map[string]interface{}, tags []string) *Execution {
+func New(workflowHash, instanceHash, parentHash, eventHash hash.Hash, taskKey string, inputs map[string]interface{}, tags []string) *Execution {
 	exec := &Execution{
+		WorkflowHash: workflowHash,
 		EventHash:    eventHash,
 		InstanceHash: instanceHash,
 		ParentHash:   parentHash,