+++ server/grpc/api/execution_test.go
@@ -20,13 +20,13 @@ func TestGet(t *testing.T) {
 	defer db.Close()
 	defer os.RemoveAll(execdbname)
 
-	exec := execution.New(nil, nil, nil, "", nil, nil)
+	exec := execution.New(nil, nil, nil, nil, "", nil, nil)
 	require.NoError(t, db.Save(exec))
 
 	want, err := toProtoExecution(exec)
 	require.NoError(t, err)
 
-	sdk := sdk.New(nil, nil, nil, db, "", "")
+	sdk := sdk.New(nil, nil, nil, db, nil, "", "")
 	s := NewExecutionServer(sdk)
 
 	got, err := s.Get(context.Background(), &api.GetExecutionRequest{Hash: exec.Hash.String()})
@@ -40,10 +40,10 @@ func TestUpdate(t *testing.T) {
 	defer db.Close()
 	defer os.RemoveAll(execdbname)
 
-	exec := execution.New(nil, nil, nil, "", nil, nil)
+	exec := execution.New(nil, nil, nil, nil, "", nil, nil)
 	require.NoError(t, db.Save(exec))
 
-	sdk := sdk.New(nil, nil, nil, db, "", "")
+	sdk := sdk.New(nil, nil, nil, db, nil, "", "")
 	s := NewExecutionServer(sdk)
 
 	_, err = s.Update(context.Background(), &api.UpdateExecutionRequest{Hash: exec.Hash.String()})